+++ packages/database/prisma/migrations/20231011213532_add_auto_yield_threshold_settings_to_endpoints/migration.sql
@@ -1,0 +1,5 @@
+-- AlterTable
+ALTER TABLE "Endpoint" ADD COLUMN     "afterCompleteTaskThreshold" INTEGER NOT NULL DEFAULT 750,
+ADD COLUMN     "beforeCompleteTaskThreshold" INTEGER NOT NULL DEFAULT 750,
+ADD COLUMN     "beforeExecuteTaskThreshold" INTEGER NOT NULL DEFAULT 1500,
+ADD COLUMN     "startTaskThreshold" INTEGER NOT NULL DEFAULT 750;