+++ integrations/openai/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @trigger.dev/slack
 
+## 2.2.2
+
+### Patch Changes
+
+- 36a39044: Upgrade to openai-node SDK 4.13.0, add support for request options and defaultQuery/defaultHeaders
+  - @trigger.dev/integration-kit@2.2.2
+  - @trigger.dev/sdk@2.2.2
+
+## 2.2.1
+
+### Patch Changes
+
+- 07d28ee4: Add ability to set baseURL to use different APIs (like Perplexity)
+- Updated dependencies [044d38e3]
+  - @trigger.dev/sdk@2.2.1
+  - @trigger.dev/integration-kit@2.2.1
+
 ## 2.2.0
 
 ### Patch Changes