+++ packages/react/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @trigger.dev/react
 
+## 2.2.2
+
+### Patch Changes
+
+- @trigger.dev/core@2.2.2
+
+## 2.2.1
+
+### Patch Changes
+
+- Updated dependencies [044d38e3]
+- Updated dependencies [abc9737a]
+  - @trigger.dev/core@2.2.1
+
 ## 2.2.0
 
 ### Patch Changes